rename: PopupPreviewResponseDTO 파일 분리
