+++ Popcorn-iOS/Source/Data/DTO/MainScene/PopupMainListResponseDTO.swift
@@ -5,6 +5,8 @@
 //  Created by 제민우 on 2/5/25.
 //
 
+import Foundation
+
 struct PopupMainListResponseDTO: Decodable {
     let userPickPopups: [PopupPreviewResponseDTO]
     let userInterestPopups: [String: PopupPreviewResponseDTO]
@@ -15,13 +17,4 @@ struct PopupMainListResponseDTO: Decodable {
         case userInterestPopups = "categoryPopups"
         case closingSoonPopups = "allPopups"
     }
-}
-
-struct PopupPreviewResponseDTO: Decodable {
-    let popupId: Int
-    let title: String
-    let imageUrl: String
-    let startDate: String
-    let endDate: String
-    let address: String
 }
